+++ src/app/home/home.component.ts
@@ -30,8 +30,12 @@ export class HomeComponent {
   filteredLocationList: HousingLocation[] = [];
 
   constructor() {
-    this.housingLocationList = this.housingService.getAllHousingLocations();
-    this.filteredLocationList = this.housingLocationList;
+    this.housingService
+      .getAllHousingLocations()
+      .then((housingLocationList: HousingLocation[]) => {
+        this.housingLocationList = housingLocationList;
+        this.filteredLocationList = housingLocationList;
+      });
   }
 
   filterResults(text: string) {
@@ -39,7 +43,6 @@ export class HomeComponent {
       this.filteredLocationList = this.housingLocationList;
       return;
     }
-
     this.filteredLocationList = this.housingLocationList.filter((housingLocation) =>
       housingLocation?.city
         .toLowerCase()